Add stock added as expenses

+++ app/src/main/java/com/vag/lmsapp/app/joborders/create/shared_ui/CreateJobOrderModifyQuantityBottomSheetFragment.kt
@@ -52,7 +52,7 @@ class CreateJobOrderModifyQuantityBottomSheetFragment : ModalFragment<QuantityMo
             binding.buttonMinus.isEnabled = true
             binding.textQuantity.apply {
                 setText(
-                    (text.toString().toInt().plus(1)).toString()
+                    (text.toString().toFloat().plus(1)).toString()
                 )
             }
         }
